refactor: migrate report chart call-sites to shared Atlas* components
Repoint PieChart/BarChart imports in ObservationPeriodReport, DeathReport,
PersonReport, and DrilldownDetails from old reports/charts/* components to
AtlasPieChart/AtlasBarChart. Migrate DrilldownDetails LineChart to
AtlasLineChart with x-axis-type="time" and reconstructed monthCodes from
MM/YYYY date strings.
