+++ src/components/datasources/DeathReport.vue
@@ -91,7 +91,7 @@ import { useI18n } from '@/composables/useI18n'
 import type { DeathReport } from '@/models/datasource.types'
 import ChartSection from './shared/ChartSection.vue'
 import EmptyReportState from './shared/EmptyReportState.vue'
-import PieChart from '@/components/reports/charts/PieChart.vue'
+import PieChart from '@/components/ui/charts/AtlasPieChart.vue'
 import BoxPlotChart from '@/components/ui/charts/AtlasBoxPlotChart.vue'
 import TrellisChart from '@/components/ui/charts/AtlasTrellisChart.vue'
 import MultiLineChart from './charts/MultiLineChart.vue'
