+++ src/components/datasources/ObservationPeriodReport.vue
@@ -160,9 +160,9 @@ import { useI18n } from '@/composables/useI18n'
 import type { ObservationPeriodReport } from '@/models/datasource.types'
 import ChartSection from './shared/ChartSection.vue'
 import EmptyReportState from './shared/EmptyReportState.vue'
-import BarChart from '@/components/reports/charts/BarChart.vue'
+import BarChart from '@/components/ui/charts/AtlasBarChart.vue'
 import BoxPlotChart from '@/components/ui/charts/AtlasBoxPlotChart.vue'
-import PieChart from '@/components/reports/charts/PieChart.vue'
+import PieChart from '@/components/ui/charts/AtlasPieChart.vue'
 import MultiLineChart from './charts/MultiLineChart.vue'
 import DashboardAgeChart from './charts/DashboardAgeChart.vue'
